backend/requirements.txt
flask==2.1.1
werkzeug==2.0.3
python-dotenv==0.19.0
openai==0.28
google-generativeai>=0.3.2,<1.0

# Audio processing
librosa
soundfile
numpy
pydub
scipy

# NLP
nltk

